main/requirements.txt
fastapi
uvicorn
mcp
fastmcp
langgraph
langchain-anthropic
pymongo
pydantic
matplotlib
seaborn
python-dotenv
pandas
